+++ src/misc.rs
@@ -1,41 +1,5 @@
-// misc.rs
-use crate::{utils::Images, users::*};
 use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Review {
-    pub mal_id: i32,
-    pub url: String,
-    #[serde(rename = "type")]
-    pub type_: String,
-    pub reactions: Option<ReviewReactions>,
-    pub date: String,
-    pub review: String,
-    pub score: i32,
-    pub tags: Vec<String>,
-    pub is_spoiler: bool,
-    pub is_preliminary: bool,
-    pub episodes_watched: Option<i32>,
-    pub user: ReviewUser,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ReviewReactions {
-    pub overall: i32,
-    pub nice: i32,
-    pub love_it: i32,
-    pub funny: i32,
-    pub confusing: i32,
-    pub informative: i32,
-    pub well_written: i32,
-    pub creative: i32,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExternalLink {
-    pub name: String,
-    pub url: String,
-}
+use crate::utils::Images;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Relation {
@@ -88,14 +52,6 @@ pub struct ImageFormat {
     pub image_url: String,
     pub small_image_url: String,
     pub large_image_url: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum ForumFilter {
-    All,
-    Episode,
-    Other,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
