+++ src/users.rs
@@ -129,16 +129,6 @@ pub struct Friend {
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct UserUpdate {
-    pub user: User,
-    pub score: Option<i32>,
-    pub status: String,
-    pub chapters_read: Option<i32>,
-    pub chapters_total: Option<i32>,
-    pub date: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct UserHistory {
     pub entry: RelatedEntry,
     pub increment: Option<i32>,
